@@ -30,11 +30,8 @@ module.exports = {
     plugins: [
         new webpack.DefinePlugin({
             PRESIDIUM_VERSION: JSON.stringify(process.env.npm_package_version),
-        }),
-        new webpack.DefinePlugin({
-            'process.env': {
-                'NODE_ENV': '"production"'
-            }
+            'process.env.NODE_ENV': JSON.stringify('production')
+
         }),
         new UglifyJsPlugin()
     ]
